requirements.txt
# Core
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Data processing
pybaseball>=2.2.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# ML (Phase 2, install now so it's ready)
scikit-learn>=1.3.0
xgboost>=2.0.0

# Dashboard (Phase 5, install later)
# streamlit>=1.29.0

# Experiment tracking
# mlflow>=2.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0